Convert onStartup into regular function calls
