+++ packages/lesswrong/client/autoRefresh.ts
@@ -1,4 +1,4 @@
-import { onStartup, getWebsocketPort } from '../lib/executionEnvironment';
+import { getWebsocketPort } from '../lib/executionEnvironment';
 import type { MessageEvent, OpenEvent, CloseEvent } from 'ws';
 
 // In development, make a websocket connection (on a different port) to get
@@ -59,8 +59,8 @@ function disconnectWebsocket() {
   }
 }
 
-if (!bundleIsProduction) {
-  onStartup(() => {
+export function initAutoRefresh() {
+  if (!bundleIsProduction) {
     setTimeout(() => {
       connectWebsocket();
       
@@ -84,5 +84,5 @@ if (!bundleIsProduction) {
         connectWebsocket();
       }
     });
-  });
+  }
 }
