+++ packages/lesswrong/client/start.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import AppGenerator from './AppGenerator';
-import { onStartup } from '../lib/executionEnvironment';
 
 import { createApolloClient } from './apolloClient';
 import { fmCrosspostBaseUrlSetting } from "../lib/instanceSettings";
@@ -8,7 +7,7 @@ import { populateComponentsAppDebug } from '../lib/vulcan-lib';
 import { initServerSentEvents } from "./serverSentEventsClient";
 import { hydrateRoot } from 'react-dom/client';
 
-onStartup(() => {
+export function hydrateClient() {
   populateComponentsAppDebug();
   initServerSentEvents();
   const apolloClient = createApolloClient();
@@ -41,5 +40,4 @@ onStartup(() => {
     apolloClient.disableNetworkFetches = false;
     foreignApolloClient.disableNetworkFetches = false;
   });
-// Order 100 to make this execute last
-}, 100);
+};
