+++ packages/lesswrong/lib/executionEnvironment.ts
@@ -29,31 +29,6 @@ export interface CommandLineArguments {
   command?: string,
 }
 
-let alreadyRunStartupFuntions = false
-
-type StartupFunction = {
-  fn: () => void|Promise<void>,
-  order: number
-}
-const onStartupFunctions: StartupFunction[] = [];
-// Register a function to be executed on startup (after top-level import is
-// done). Startup functions have a numeric order attached, and are executed in
-// order from lowest to highest. If no order is given, the order is 0. Between
-// functions with the same order number, order of execution is undefined.
-export const onStartup = (fn: () => void|Promise<void>, order?: number) => {
-  if (alreadyRunStartupFuntions) {
-    throw new Error("Startup functions have already been run, can no longer register more")
-  }
-  onStartupFunctions.push({fn, order: order||0});
-}
-
-export const runStartupFunctions = async () => {
-  alreadyRunStartupFuntions = true
-  for (let startupFunction of _.sortBy(onStartupFunctions, f=>f.order)) {
-    await startupFunction.fn();
-  }
-}
-
 let instanceSettings: any = null;
 export const getInstanceSettings = (args?: CommandLineArguments): any => {
   if (!instanceSettings) {
