+++ packages/lesswrong/server/cronUtil.ts
@@ -1,4 +1,4 @@
-import { isAnyTest, isDevelopment, onStartup } from '../lib/executionEnvironment';
+import { isAnyTest, isDevelopment } from '../lib/executionEnvironment';
 import { SyncedCron } from './vendor/synced-cron/synced-cron-server';
 import { getCommandLineArguments } from './commandLine';
 import { CronHistories } from '../lib/collections/cronHistories';
@@ -20,27 +20,25 @@ export type CronJobSpec = {
 }
 
 export function addCronJob(options: CronJobSpec) {
-  onStartup(function() {
-    if (!isAnyTest && !getCommandLineArguments().shellMode) {
-      // Defer starting of cronjobs until 20s after server startup
-      setTimeout(() => {
-        SyncedCron.add({
-          name: options.name,
-          schedule: (parser: any) => {
-            if (options.interval)
-              return parser.text(options.interval);
-            else if (options.cronStyleSchedule) {
-              const hasSeconds = options.cronStyleSchedule.split(' ').length > 5;
-              return parser.cron(options.cronStyleSchedule, hasSeconds);
-            }
-            else
-              throw new Error("addCronJob needs a schedule specified");
-          },
-          job: options.job,
-        });
-      }, 20000);
-    }
-  });
+  if (!isAnyTest && !getCommandLineArguments().shellMode) {
+    // Defer starting of cronjobs until 20s after server startup
+    setTimeout(() => {
+      SyncedCron.add({
+        name: options.name,
+        schedule: (parser: any) => {
+          if (options.interval)
+            return parser.text(options.interval);
+          else if (options.cronStyleSchedule) {
+            const hasSeconds = options.cronStyleSchedule.split(' ').length > 5;
+            return parser.cron(options.cronStyleSchedule, hasSeconds);
+          }
+          else
+            throw new Error("addCronJob needs a schedule specified");
+        },
+        job: options.job,
+      });
+    }, 20000);
+  }
 }
 
 export function removeCronJob(name: string) {
@@ -52,10 +50,6 @@ export function startSyncedCron() {
     SyncedCron.start();
   }
 }
-
-onStartup(function() {
-  startSyncedCron();
-});
 
 async function clearOldCronHistories() {
   const ONE_WEEK = 7 * 24 * 60 * 60 * 1000;
