+++ packages/lesswrong/server/editor/make_editable_callbacks.ts
@@ -14,7 +14,6 @@ import {getCollection} from '../../lib/vulcan-lib/getCollection'
 import {CallbackHook} from '../../lib/vulcan-lib/callbacks'
 import {createMutator, validateCreateMutation} from '../vulcan-lib/mutators'
 import * as _ from 'underscore'
-import {onStartup} from '../../lib/executionEnvironment'
 import {dataToHTML, dataToWordCount} from './conversionUtils'
 import {Globals} from '../../lib/vulcan-lib/config'
 import {notifyUsersAboutMentions, PingbackDocumentPartial} from './mentions-notify'
@@ -342,5 +341,3 @@ export function addAllEditableCallbacks() {
     }
   }
 }
-
-onStartup(addAllEditableCallbacks);
