+++ packages/lesswrong/server/serverStartup.ts
@@ -4,20 +4,14 @@ import { createSqlConnection } from './sqlConnection';
 import { replaceDbNameInPgConnectionString, setSqlClient } from '../lib/sql/sqlClient';
 import PgCollection, { DbTarget } from '../lib/sql/PgCollection';
 import { Collections } from '../lib/vulcan-lib/getCollection';
-import { runStartupFunctions, isAnyTest, isMigrations, CommandLineArguments } from '../lib/executionEnvironment';
+import { isAnyTest, CommandLineArguments } from '../lib/executionEnvironment';
 import { PublicInstanceSetting } from "../lib/instanceSettings";
 import { refreshSettingsCaches } from './loadDatabaseSettings';
 import { getCommandLineArguments } from './commandLine';
-import { startWebserver } from './apolloServer';
-import { initGraphQL } from './vulcan-lib/apollo-server/initGraphQL';
-import { createVoteableUnionType } from './votingGraphQL';
-import { Globals, Vulcan } from '../lib/vulcan-lib/config';
+import { Globals } from '../lib/vulcan-lib/config';
 import { getBranchDbName } from "./branchDb";
 import { dropAndCreatePg } from './testingSqlClient';
 import process from 'process';
-import chokidar from 'chokidar';
-import fs from 'fs';
-import { basename, join } from 'path';
 import { filterConsoleLogSpam, wrapConsoleLogFunctions } from '../lib/consoleFilters';
 import { getAllPostgresViews } from './postgresView';
 import cluster from 'node:cluster';
@@ -114,27 +108,6 @@ const initPostgres = async () => {
   }
 }
 
-const executeServerWithArgs = async ({shellMode, command}: CommandLineArguments) => {
-  await runStartupFunctions();
-
-  // define executableSchema
-  createVoteableUnionType();
-  initGraphQL();
-
-  if (shellMode) {
-    initShell();
-  } else if (command) {
-    const func = compileWithGlobals(command);
-    const result = await func();
-    // eslint-disable-next-line no-console
-    console.log("Finished. Result: ", result);
-    process.kill(estrellaPid, 'SIGQUIT');
-  } else if (!isAnyTest && !isMigrations) {
-    watchForShellCommands();
-    startWebserver();
-  }
-}
-
 export const initServer = async (commandLineArguments?: CommandLineArguments) => {
   initConsole();
   const args = commandLineArguments ?? getCommandLineArguments();
@@ -185,75 +158,6 @@ export const serverStartup = async () => {
 
 export const serverStartupWorker = async () => {
   const commandLineArguments = await initServer();
-  await executeServerWithArgs(commandLineArguments);
-}
-
-function initShell() {
-  const repl = require('repl');
-  /*const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-    prompt: "> ",
-  });
-  rl.on('line', line => {
-    console.log(`Got input: ${line}`);
-    rl.prompt();
-  });
-  rl.prompt();*/
-
-  const r = repl.start({
-    prompt: "> ",
-    terminal: true,
-    preview: true,
-    breakEvalOnSigint: true,
-    useGlobal: true,
-  });
-  r.context.Globals = Globals;
-  r.context.Vulcan = Globals;
-}
-
-const compileWithGlobals = (code: string) => {
-  // This is basically just eval() but done in a way that:
-  //   1) Allows us to define our own global scope
-  //   2) Doesn't upset esbuild
-  const callable = (async function () {}).constructor(`with(this) { await ${code} }`);
-  const scope = {Globals, Vulcan};
-  return () => {
-    return callable.call(new Proxy({}, {
-      has () { return true; },
-      get (_target, key) {
-        if (typeof key !== "symbol") {
-          return global[key as keyof Global] ?? scope[key as "Globals"|"Vulcan"];
-        }
-      }
-    }));
-  }
-}
-
-// Monitor ./tmp/pendingShellCommands for shell commands. If a JS file is
-// written there, run it then delete it. Security-wise this is okay because
-// write-access inside the repo directory is already equivalent to script
-// execution.
-const watchForShellCommands = () => {
-  const watcher = chokidar.watch('./tmp/pendingShellCommands');
-  watcher.on('add', async (path) => {
-    const fileContents = fs.readFileSync(path, 'utf8');
-    // eslint-disable-next-line no-console
-    console.log(`Running shell command: ${fileContents}`);
-    const newPath = join("tmp/runningShellCommands", basename(path));
-    fs.renameSync(path, newPath);
-    try {
-      const func = compileWithGlobals(fileContents);
-      const result = await func();
-      // eslint-disable-next-line no-console
-      console.log("Finished. Result: ", result);
-    } catch(e) {
-      // eslint-disable-next-line no-console
-      console.log("Failed.");
-      // eslint-disable-next-line no-console
-      console.log(e);
-    } finally {
-      fs.unlinkSync(newPath);
-    }
-  });
+  const { serverMain } = require('./serverMain');
+  await serverMain(commandLineArguments);
 }
